Edited bootstrap css code and split archiving function from Delete function

+++ models.py
@@ -9,7 +9,7 @@ class Forward(db.Model):
     methode = db.Column(db.String(32))
     provision = db.Column(db.String(32), default='present')
     archive = db.Column(db.String(32), default='n')
-    timestamp = db.Column(db.Float)
+    timestamp = db.Column(db.Integer)
 
 
 class ForwardSchema(ma.ModelSchema):
